requirements.txt
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
yfinance>=0.2.0
statsmodels>=0.14.0
pykalman>=0.9.5
hmmlearn>=0.3.0
quantstats>=0.0.59
numba>=0.57.0
# Additional useful libraries
scikit-learn>=1.0.0  # For machine learning models and feature engineering
ta>=0.10.0  # Technical analysis indicators (pure Python implementation)
tqdm>=4.65.0  # Progress bars for long-running operations
PyYAML>=6.0
